feat: refactors page logic, adds context, global controlpanel
