--- src/components/_NextPages.js
@@ -1,5 +1,0 @@
-module.exports = [
-  'population',
-  'workers-field-of-qualification',
-  'value-of-building-approvals'
-];
