+++ src/pages/[clientAlias]/[handle]/index.tsx
@@ -1,0 +1,80 @@
+import { Context } from '../../../utils/context';
+
+import fetchClientData from '../../../utils/fetchClientData';
+import fetchSitemap from '../../../utils/fetchSitemap';
+import fetchIndustries from '../../../utils/fetchIndustries';
+import fetchBenchmarkAreas from '../../../utils/fetchBenchmarkAreas';
+
+import Population from '../../../layouts/population/page';
+import ValueOfBuildingApprovals from '../../../layouts/value-of-building-approvals/page';
+import WorkersFieldOfQualification from '../../../layouts/workers-field-of-qualification/page';
+
+const defaultFilters = {
+  Indkey: 23000,
+  IGBMID: 40,
+  Sex: 3,
+  WebID: 10,
+};
+
+export const NextPages = {
+  population: Population,
+  'value-of-building-approvals': ValueOfBuildingApprovals,
+  'workers-field-of-qualification': WorkersFieldOfQualification,
+};
+
+const Page = props => {
+  const Layout = NextPages[props.handle];
+
+  return (
+    <Context.Provider value={props}>
+      <Layout />
+    </Context.Provider>
+  );
+};
+
+Page.getInitialProps = async function(context) {
+  const filters = {
+    ...defaultFilters,
+    ...context.query,
+  };
+
+  const { clientAlias, handle } = context.query;
+
+  const { fetchData } = await import(`../../../layouts/${handle}`);
+
+  const clientData = await fetchClientData({
+    clientAlias,
+    containers: context.req.containers,
+  });
+
+  const { ClientID, clientAreas, clientProducts, clientPages } = clientData;
+
+  const tableData = await fetchData({ ClientID, ...filters });
+
+  const Industries = await fetchIndustries(40);
+  const BenchmarkAreas = await fetchBenchmarkAreas(ClientID);
+  const sitemapGroups = await fetchSitemap();
+
+  const Sexes = [
+    { ID: 1, Name: 'Males' },
+    { ID: 2, Name: 'Females' },
+    { ID: 3, Name: 'Persons' },
+  ];
+
+  return {
+    tableData,
+    Industries,
+    BenchmarkAreas,
+    Sexes,
+    navigation: clientPages,
+    clientProducts,
+    sitemapGroups,
+    filters,
+    clientAreas,
+    handle,
+    clientData,
+    clientAlias,
+  };
+};
+
+export default Page;
