+++ src/utils/context/index.ts
@@ -1,0 +1,16 @@
+import React from 'react';
+
+export const Context = React.createContext({
+  clientData: null,
+  clientAlias: null,
+  handle: null,
+  tableData: null,
+  navigation: null,
+  clientProducts: null,
+  sitemapGroups: null,
+  BenchmarkAreas: null,
+  Industries: null,
+  Sexes: null,
+  filters: null,
+  clientAreas: null,
+});
